requirements.txt
sanic==22.6.0
Sanic-Cors==2.1.0
sanic-routing==22.3.0
revChatGPT
python-dotenv